requirements.txt
bitalino==0.6
numpy==1.11.0
pyOSC==0.3.5b5294
pyserial==2.7
wheel==0.24.0
